@@ -25,69 +38,86 @@
     "grunt-contrib-concat": "^1.0.1",
     "grunt-contrib-copy": "~1.0.0",
     "grunt-contrib-cssmin": "~1.0.2",
-    "grunt-contrib-htmlmin": "~2.0.0",
     "grunt-contrib-jshint": "~1.1.0",
-    "grunt-contrib-uglify": "~2.0.0",
-    "grunt-contrib-watch": "^1.0.0",
     "grunt-exec": "^1.0.1",
-    "grunt-filerev": "^2.3.1",
+    "grunt-jscs": "3.0.1",
     "grunt-karma": "~2.0.0",
-    "grunt-ng-annotate": "^3.0.0",
     "grunt-notify": "^0.4.5",
     "grunt-postcss": "^0.8.0",
     "grunt-sass": "^2.0.0",
-    "grunt-string-replace": "~1.3.1",
+    "grunt-sass-lint": "^0.2.2",
     "grunt-usemin": "3.1.1",
+    "grunt-webpack": "^3.0.2",
+    "html-loader": "^0.5.1",
+    "html-webpack-plugin": "^2.30.1",
     "jshint-stylish": "~2.2.1",
+    "json-loader": "^0.5.7",
     "karma": "1.7.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage": "1.1.1",
     "karma-expect": "~1.1.3",
     "karma-mocha": "~1.3.0",
     "karma-phantomjs-launcher": "1.0.4",
+    "karma-sinon": "^1.0.5",
+    "karma-sourcemap-loader": "^0.3.7",
+    "karma-webpack": "^2.0.4",
     "load-grunt-tasks": "3.5.2",
     "mocha": "3.5.0",
+    "ng-annotate-loader": "^0.6.1",
+    "ng-annotate-webpack-plugin": "^0.2.1-pre",
+    "ngtemplate-loader": "^2.0.1",
     "phantomjs-prebuilt": "^2.1.15",
+    "postcss-browser-reporter": "^0.5.0",
+    "postcss-loader": "^2.0.6",
+    "postcss-reporter": "^5.0.0",
+    "react-test-renderer": "^16.0.0",
     "rxjs": "^5.4.3",
     "sass-lint": "^1.10.2",
-    "systemjs": "0.20.19"
+    "sass-loader": "^6.0.6",
+    "sinon": "1.17.6",
+    "systemjs": "0.20.19",
+    "systemjs-plugin-css": "^0.1.36",
+    "ts-loader": "^2.3.7",
+    "tslint": "^5.7.0",
+    "tslint-loader": "^3.5.3",
+    "typescript": "^2.5.2",
+    "webpack": "^3.6.0",
+    "webpack-bundle-analyzer": "^2.9.0",
+    "webpack-cleanup-plugin": "^0.5.1",
+    "webpack-merge": "^4.1.0",
+    "zone.js": "^0.7.2",
+    "awesome-typescript-loader": "^3.2.3",
+    "angular-mocks": "^1.6.6",
+    "npm": "^5.4.2"
   },
   "scripts": {
+    "webpack": "./node_modules/.bin/webpack --progress --colors --config scripts/webpack/webpack.dev.js",
+    "watch": "./node_modules/.bin/webpack --progress --colors --watch --config scripts/webpack/webpack.dev.js",
     "build": "./node_modules/grunt-cli/bin/grunt",
     "test": "./node_modules/grunt-cli/bin/grunt test",
-    "dev": "./node_modules/grunt-cli/bin/grunt && ./node_modules/grunt-cli/bin/grunt watch"
+    "watch-test": "./node_modules/grunt-cli/bin/grunt karma:dev"
   },
   "license": "Apache-2.0",
   "dependencies": {
-    "@types/enzyme": "^2.8.8",
-    "ace-builds": "^1.2.8",
+    "babel-polyfill": "^6.26.0",
+    "jquery": "^3.2.1",
     "angular": "^1.6.6",
     "angular-bindonce": "^0.3.1",
-    "angular-mocks": "^1.6.6",
     "angular-native-dragdrop": "^1.2.2",
     "angular-route": "^1.6.6",
     "angular-sanitize": "^1.6.6",
+    "brace": "^0.10.0",
     "clipboard": "^1.7.1",
-    "eventemitter3": "^2.0.2",
-    "gaze": "^1.1.2",
-    "grunt-jscs": "3.0.1",
-    "grunt-sass-lint": "^0.2.2",
-    "grunt-sync": "^0.6.2",
-    "jquery": "^3.2.1",
-    "karma-sinon": "^1.0.5",
+    "eventemitter3": "^2.0.3",
+    "file-saver": "^1.3.3",
     "lodash": "^4.17.4",
+    "moment": "^2.18.1",
     "mousetrap": "^1.6.0",
     "ngreact": "^0.4.1",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
-    "react-test-renderer": "^15.6.1",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
     "remarkable": "^1.7.1",
-    "sinon": "1.17.6",
-    "systemjs-builder": "^0.16.10",
     "tether": "^1.4.0",
-    "tether-drop": "https://github.com/torkelo/drop",
-    "tslint": "^5.7.0",
-    "typescript": "^2.5.2",
-    "virtual-scroll": "^1.1.1"
+    "tether-drop": "https://github.com/torkelo/drop"
   }
 }
