@@ -3,7 +3,7 @@
 import _ from 'lodash';
 import config from 'app/core/config';
 
-import {Observable} from 'vendor/npm/rxjs/Observable';
+import {Observable} from 'rxjs/Observable';
 
 export class LiveSrv {
   conn: any;
