@@ -1,23 +1,31 @@
+var webpack = require('webpack');
+var path = require('path');
+var webpackTestConfig = require('./scripts/webpack/webpack.test.js');
+
 module.exports = function(config) {
+
   'use strict';
 
   config.set({
-    basePath: __dirname + '/public_gen',
-
     frameworks: ['mocha', 'expect', 'sinon'],
 
     // list of files / patterns to load in the browser
     files: [
-      'vendor/npm/es6-shim/es6-shim.js',
-      'vendor/npm/systemjs/dist/system.src.js',
-      'test/test-main.js',
-
-      {pattern: '**/*.js', included: false},
+      'public/test/index.ts',
     ],
+
+    preprocessors: {
+      'public/test/index.ts': ['webpack', 'sourcemap'],
+    },
+
+    webpack: webpackTestConfig,
+
+    webpackServer: {
+      noInfo: true, // please don't spam the console when running in karma!
+    },
 
     // list of files to exclude
     exclude: [],
-
     reporters: ['dots'],
     port: 9876,
     colors: true,
@@ -26,8 +34,8 @@ module.exports = function(config) {
     browsers: ['PhantomJS'],
     captureTimeout: 20000,
     singleRun: true,
-    autoWatchBatchDelay: 1000,
-    browserNoActivityTimeout: 60000,
+    // autoWatchBatchDelay: 1000,
+    // browserNoActivityTimeout: 60000,
   });
 
 };
