@@ -15,14 +15,14 @@ export class PasswordStrength extends React.Component<IProps, any> {
     let strengthText = "strength: strong like a bull.";
     let strengthClass = "password-strength-good";
 
-    if (this.props.password.length < 4) {
-      strengthText = "strength: weak sauce.";
-      strengthClass = "password-strength-bad";
-    }
-
     if (this.props.password.length <= 8) {
       strengthText = "strength: you can do better.";
       strengthClass = "password-strength-ok";
+    }
+
+    if (this.props.password.length < 4) {
+      strengthText = "strength: weak sauce.";
+      strengthClass = "password-strength-bad";
     }
 
     return (
