@@ -1,7 +1,4 @@
-///<reference path="../../headers/common.d.ts" />
-
 import $ from 'jquery';
-
 import coreModule from '../core_module';
 
 function getBlockNodes(nodes) {
@@ -21,6 +18,7 @@ function getBlockNodes(nodes) {
   return blockNodes || nodes;
 }
 
+/** @ngInject **/
 function rebuildOnChange($animate) {
 
   return {
