@@ -1,6 +1,4 @@
-///<reference path="../../../headers/common.d.ts" />
-
-import 'jquery.flot';
+import 'vendor/flot/jquery.flot';
 import $ from 'jquery';
 import _ from 'lodash';
 
