@@ -1,8 +1,6 @@
-///<reference path="../../../../headers/common.d.ts" />
+import {describe, beforeEach, afterEach, it, sinon, expect, angularMocks} from 'test/lib/common';
 
-import {describe, beforeEach, afterEach, it, sinon, expect, angularMocks} from '../../../../../test/lib/common';
-
-import helpers from '../../../../../test/specs/helpers';
+import helpers from 'test/specs/helpers';
 import {SingleStatCtrl} from '../module';
 import moment from 'moment';
 
