@@ -1,6 +1,5 @@
-///<reference path="../../../headers/common.d.ts" />
 
-import {PanelCtrl} from 'app/plugins/sdk';
+import {PanelCtrl} from 'app/features/panel/panel_ctrl';
 
 export class UnknownPanelCtrl extends PanelCtrl {
   static templateUrl = 'public/app/plugins/panel/unknown/module.html';
@@ -9,7 +8,6 @@ export class UnknownPanelCtrl extends PanelCtrl {
   constructor($scope, $injector) {
     super($scope, $injector);
   }
-
 }
 
 
