@@ -2,20 +2,12 @@ import {PanelCtrl} from 'app/features/panel/panel_ctrl';
 import {MetricsPanelCtrl} from 'app/features/panel/metrics_panel_ctrl';
 import {QueryCtrl} from 'app/features/panel/query_ctrl';
 import {alertTab} from 'app/features/alerting/alert_tab_ctrl';
-
-import config from 'app/core/config';
-
-export function loadPluginCss(options) {
-  if (config.bootData.user.lightTheme) {
-    System.import(options.light + '!css');
-  } else {
-    System.import(options.dark + '!css');
-  }
-}
+import {loadPluginCss} from 'app/features/plugins/plugin_loader';
 
 export {
   PanelCtrl,
   MetricsPanelCtrl,
   QueryCtrl,
   alertTab,
+  loadPluginCss,
 };
